#12: moved db.User to more correct authuserdao.User
